ZRAM (1GB) still needed in a 8GB RAM device... ?
This is not a real change proposition (although I would like to avoid ZRAM management because I think it's an unnecesary overload...). Only change it if you also think it worths...
I would like to discuss if ZRAM is still needed nowdays... I see many kernels in modern devices implement it, but I consider that maybe only because it's a "default" setting...
Thanks in advance.